+++ src/data/resouce/local/rating_score/RatingScoreLocalDataSourceImpl.kt
@@ -42,7 +42,7 @@ class RatingScoreLocalDataSourceImpl(
         ratingScoreId: String,
         userId: String,
         foodId: Int,
-        ratingScore: Int,
+        ratingScore: Float,
         isBackup: Int,
         created: String,
         updated: String?
