+++ src/route/models/response/RatingScoreResponse.kt
@@ -5,7 +5,7 @@ data class RatingScoreResponse(
     val ratingScoreId: String,
     val userId: String,
     val foodId: Int,
-    val ratingScore: Int,
+    val ratingScore: Float,
     val isBackup: Boolean,
     val created: String,
     val updated: String?,
